refactor: Admin page 컴포넌트 분리
